Switches to absolute voxel positions (ivec).

+++ src/BoxVoxelData.h
@@ -10,8 +10,8 @@ using namespace glm;
  */
 class BoxVoxelData : public VoxelData {
  public:
- BoxVoxelData(vec3 corner, vec3 dim , Voxel v) : c1_(corner), c2_(corner + dim), v_(v) { };
-  virtual bool voxelAt(vec3 position, Voxel* v) {
+ BoxVoxelData(ivec3 corner, ivec3 dim , Voxel v) : c1_(corner), c2_(corner + dim), v_(v) { };
+  virtual bool voxelAt(ivec3 position, Voxel* v) {
     if ( position.x < c2_.x && position.x >= c1_.x &&
 	 position.y < c2_.y && position.y >= c1_.y &&
 	 position.z < c2_.z && position.z >= c1_.z ) {
@@ -22,6 +22,6 @@ class BoxVoxelData : public VoxelData {
     }
   };
  private:
-  vec3 c1_, c2_; // Dimensions
+  ivec3 c1_, c2_; // Dimensions
   Voxel v_; // Color
 };
